portfolio 102, worked on portfolio-100, fixed, optimized 90%
