+++ portfolio/src/app/app.component.html
@@ -12,6 +12,12 @@
    </div> 
    -->
 
+   <!--   at SizeOnlySource._error (F:\Projects\PortfolioProject\portfolio\node_modules\webpack-sources\lib\SizeOnlySource.js:16:10)
+          npm install source-map-support
+          npm install
+   
+   -->
+
    <div class = "site-container">
       <div id = "siteContainerTop" class = "site-container-top">
          <div class = 'view-container'>
